extract statistics and account data to separate hooks
